Changed the file name of the start.gg functions to startgg_utils
