+++ src/get_participants/mapping.py
@@ -1,12 +1,12 @@
 from discord import AppCommandOptionType
 from commands.models.command_mapping import CommandMapping
 from commands.models.command_param import CommandParam
-from get_participants.startgg.get_startgg_participants_main import get_startgg_participants_list
+from get_participants.startgg.get_startgg_participants import get_startgg_participants_msg
 
 
 get_participants_commands: CommandMapping = {
     "get_startgg_participants": {
-        "function": get_startgg_participants_list,
+        "function": get_startgg_participants_msg,
         "description": "Retrieves list of participants of an event",
         "params": [
             CommandParam(name = "bracket_link",
